RecursiveFiles

Documentation for RecursiveFiles.

RecursiveFiles.include_files

RecursiveFiles.list_files

RecursiveFiles.include_files — Method

retval = include_files(srcdir; fileToRemove="NeverToBeFound", pattern=juliaFilesPattern)

Include all Julia files recursively in folder SRCDIR,
excluding the module file which is to be specified as the FILETOREMOVE.
The optional PATTERN is not required here as it includes all Julia files per default.

RecursiveFiles.list_files — Method

filelist = list_files(srcdir; fileToRemove="NeverToBeFound", pattern=allFilesPattern)

Get a recursive list of files in folder SRCDIR (full nested paths).
The optional FILETOREMOVE (if present) will be removed from the list
The optional PATTERN can be used for filtering for certain file types.